fix: set fallback src to avoid exception (#86)
On avatar change the exception:

```
An empty string ("") was passed to the src attribute. This may cause the browser to download the whole page again over the network. To fix this, either do not render the element at all or pass null to src instead of an empty string.
```

This PR fixes this by adding default src